# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "handlers.FailureResponse": {
        "properties": {
          "attempts": {
            "type": "integer"
          },
          "error_reason": {
            "type": "string"
          },
          "error_type": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last_attempt_at": {
            "type": "string"
          },
          "source_file": {
            "type": "string"
          },
          "symbol_kind": {
            "type": "string"
          },
          "symbol_name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "handlers.RetryRequest": {
        "properties": {
          "failure_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "workspace": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.RetryResponse": {
        "properties": {
          "failed": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "retried": {
            "type": "integer"
          },
          "succeeded": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "handlers.graphEdgeItem": {
        "properties": {
          "edge_type": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.graphQueryRequest": {
        "properties": {
          "direction": {
            "type": "string"
          },
          "edge_type": {
            "type": "string"
          },
          "node": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.graphQueryResponse": {
        "properties": {
          "direction": {
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/handlers.graphEdgeItem"
            },
            "type": "array"
          },
          "node": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CSRFToken": {
        "description": "Required on write endpoints; enforced by middleware.CSRF.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "WorkspaceAuth": {
        "description": "Workspace hash via query param or JSON body; resolved by workspaceMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      },
      "WorkspaceRegisteredAuth": {
        "description": "Workspace must be already-registered (write-path gate); enforced by workspaceRegisteredMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Self-describing OpenAPI 3.0 spec for nano-brain's non-MCP REST API",
    "title": "nano-brain REST API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/code/summarize/failures": {
      "get": {
        "description": "Returns unresolved code-symbol summarization failures for the workspace",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/handlers.FailureResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceRegisteredAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "List unresolved code summarization failures",
        "tags": [
          "code-summarize"
        ]
      }
    },
    "/api/v1/code/summarize/retry": {
      "post": {
        "description": "Retries code summarization for the given failure_ids",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.RetryRequest"
              }
            }
          },
          "description": "Retry options (failure_ids required)",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.RetryResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceRegisteredAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Retry specific failed code summarizations",
        "tags": [
          "code-summarize"
        ]
      }
    },
    "/api/v1/code/summarize/retry-all": {
      "post": {
        "description": "Retries code summarization for every unresolved failure in the workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.RetryRequest"
              }
            }
          },
          "description": "Retry options",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.RetryResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceRegisteredAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Retry all failed code summarizations",
        "tags": [
          "code-summarize"
        ]
      }
    },
...
    "/api/v1/graph/query": {
      "post": {
        "description": "Returns outgoing, incoming, or both-direction edges for a given graph node, scoped to a workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.graphQueryRequest"
              }
            }
          },
          "description": "Graph query",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.graphQueryResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceAuth": []
          }
        ],
        "summary": "Query graph edges for a node",
        "tags": [
          "graph"
        ]
      }
    },
...
  }
}